Ecommerce-App/Backend/requirements.txt
Flask
Flask-Cors
python-dotenv
Flask-SQLAlchemy
psycopg2-binary